[Fix/#176] 기록 삭제 시 코루틴 잡 캔슬 오류 해결
